third_party/blink/web_tests/svg/animations/deferred-insertion-expected.txt

SVG 1.1 dynamic animation tests

Test for animation on elements inserted programmatically. Should result in a 200x200 rect and only PASS messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS rect.width.animVal.value is 200
PASS rect.width.baseVal.value is 200
PASS rect.width.animVal.value is 150
PASS rect.width.baseVal.value is 200
PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 200
PASS rect.width.animVal.value is 200
PASS rect.width.baseVal.value is 200
PASS successfullyParsed is true

TEST COMPLETE
